requirements.txt
aiohttp==3.7.4.post0
boto3==1.16.38
botocore==1.19.63
dotmap==1.3.23
discord.py==1.7.3
eip712==0.1.0
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.0
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
gql==3.0.0a5
isort==5.10.1
py==1.10.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
python-decouple==3.3
python-dotenv==0.16.0
requests==2.26.0
rich==9.3.0
tabulate==0.8.7
